+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v2.3.0
+- Added compendium token support for "Spawn riders"
+- Added "cluster" placement pattern
+- Added "Adjust rider size" setting
+
 ## v2.2.1
 - Bug fix for free riding tokens and stairways compatibility